+++ Changelog.md
@@ -6,6 +6,7 @@ Bugfixes:
  * Code Generator: ``.delegatecall()`` should always return execution outcome.
  * Code Generator: Provide "new account gas" for low-level ``callcode`` and ``delegatecall``.
  * Type Checker: Disallow the ``.gas()`` modifier on ``ecrecover``, ``sha256`` and ``ripemd160``.
+ * Type Checker: Raise error when using unimplemented internal library functions.
 
 ### 0.4.14 (2017-07-31)